Braking changes: - reexport only required types inside lib: NodeInfo and HashRing - removed example.rs. Now tests are inside hash_ring
